Adds Docker setup for easier development
Simplifies the setup process by introducing Docker Compose for running PostgreSQL in a container.

Provides `make` commands for managing the database and development server.

Updates the README with instructions for quick start using Docker, as well as manual installation.

Changes default database name to `collabotask_dev` and updates database credentials to be Docker-friendly.